# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TournaX Gaming Tournament API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, completed]
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "200":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    delete:
      operationId: deleteMatch
      tags: [matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /matches/{id}/join:
    post:
      operationId: joinMatch
      tags: [matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinMatchRequest"
      responses:
        "200":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /matches/{id}/room:
    put:
      operationId: updateRoomCredentials
      tags: [matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomCredentialsRequest"
      responses:
        "200":
          description: Room credentials updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /matches/{id}/players:
    get:
      operationId: getMatchPlayers
      tags: [matches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchParticipant"
...
  /admin/matches:
    get:
      operationId: adminListMatches
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, upcoming, live, completed]
      responses:
        "200":
          description: All matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Match:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        game:
          type: string
        mode:
          type: string
        teamSize:
          type: integer
        entryFee:
          type: number
        prizePool:
          type: number
        startTime:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        slots:
          type: integer
        filledSlots:
          type: integer
        hostHandle:
          type: string
        hostName:
          type: string
        hostAvatar:
          type: string
        hostFollowers:
          type: integer
        isFollowingHost:
          type: boolean
        isJoined:
          type: boolean
        roomId:
          type: string
        roomPassword:
          type: string
        roomReleased:
          type: boolean
        hostId:
          type: integer
      required: [id, code, game, mode, teamSize, entryFee, prizePool, startTime, status, slots, filledSlots, hostId]
    CreateMatchRequest:
      type: object
      properties:
        game:
          type: string
        mode:
          type: string
        teamSize:
          type: integer
        entryFee:
          type: number
        slots:
          type: integer
        startTime:
          type: string
      required: [game, mode, teamSize, entryFee, slots, startTime]
    JoinMatchRequest:
      type: object
      properties:
        teamName:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/JoinPlayer"
      required: [players]
    JoinPlayer:
      type: object
      properties:
        ign:
          type: string
        uid:
          type: string
      required: [ign, uid]
    MatchParticipant:
      type: object
      properties:
        id:
          type: integer
        teamName:
          type: string
        teamNumber:
          type: integer
        players:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantPlayer"
      required: [id, teamNumber, players]
    ParticipantPlayer:
      type: object
      properties:
        ign:
          type: string
        uid:
          type: string
        position:
          type: integer
      required: [ign, uid, position]
    RoomCredentialsRequest:
      type: object
      properties:
        roomId:
          type: string
        roomPassword:
          type: string
      required: [roomId, roomPassword]
...